+++ CHANGELOG.md
@@ -9,3 +9,9 @@ Manual graphics resources control support.
 
 # 2024.06.21
 Text with custom fonts support.
+
+# 2024.07.08
+Clj format animation feature.
+
+# 2024.07.14
+Custom materials support.